Send initial chunks (only blocks for now) to clients Those are decompressed using the generation thread pool on clients TODO: - Compress and send chunks using the thread pool as well, currently it   freeze the server

+++ src/Network/Packet.hpp
@@ -291,7 +291,6 @@ struct ChunkDataPacket
     int64_t x;
     int64_t z;
     Vector<uint8_t> blocks;
-    // Vector<Biome> biomes;
 
     static constexpr PacketType type = PacketType::ChunkData;
 };
@@ -304,7 +303,6 @@ inline Result<void> serialize(DataBuffer& buffer, const ChunkDataPacket& p)
     buffer.write(size);
 
     buffer.write_array(p.blocks);
-    // buffer.write_array(p.biomes);
     return Result<void>();
 }
 inline Result<void> deserialize(DataBuffer& buffer, ChunkDataPacket& p)
@@ -312,8 +310,7 @@ inline Result<void> deserialize(DataBuffer& buffer, ChunkDataPacket& p)
     p.x = buffer.read<int64_t>();
     p.z = buffer.read<int64_t>();
 
-    uint32_t size = buffer.read<uint8_t>();
+    uint32_t size = buffer.read<uint32_t>();
     p.blocks = buffer.read_array<uint8_t>(size);
-    // p.biomes = buffer.read_array<Biome>(Chunk::block_count);
     return Result<void>();
 }
